KAdaptiveAlgorithm: use float instead of int for calculating ratios.

+++ src/us/mn/state/dot/tms/server/KAdaptiveAlgorithm.java
@@ -1183,7 +1183,7 @@ public class KAdaptiveAlgorithm implements MeterAlgorithmState {
 		}
 
 		/** Get the max wait time (seconds) */
-		private int maxWaitTime() {
+		private float maxWaitTime() {
 			return Math.max(meter.getMaxWait(), 1);
 		}
